added models and repos for CIModule, RemoteControllers and WallBrackets

+++ src/main/java/nl/novi/bachwtechiteasy/repositories/TelevisionRepository.java
@@ -3,5 +3,5 @@ package nl.novi.bachwtechiteasy.repositories;
 import nl.novi.bachwtechiteasy.models.Television;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TelevisionRepository extends JpaRepository<Television, Integer> {
+public interface TelevisionRepository extends JpaRepository<Television, Long> {
 }
